+++ examples/complex/global/boot.rb
@@ -6,3 +6,7 @@
 require 'dotenv/load'
 
 puts "Booting #{Config::SHOP_NAME}..."
+
+# Add extra root-level autoload dirs via Boxwerk.global.autoloader.
+# Constants loaded here are available in all package boxes.
+Boxwerk.global.autoloader.push_dir(File.expand_path('../lib', __dir__))